Refactor BlockchainCheckpoints code for easier readability
- Also so we can use it in process_block(), without duplicating entire check code
